+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "pretest": "eslint index.js lib test/*.js",
-    "test": "nyc mocha test/*.test.js",
-    "coverage": "nyc --reporter html mocha test/*.test.js && opener ./coverage/index.html",
+    "test": "nyc tape test/*.test.js",
+    "coverage": "nyc --reporter html tape test/*.test.js && opener ./coverage/index.html",
     "docs": "documentation -f md --github > api.md",
     "posttest": "rm -r ./test/test"
   },
@@ -52,8 +52,8 @@
     "geojson-fixtures": "0.1.0",
     "geojson-random": "^0.2.2",
     "geojson-stream": "0.0.0",
-    "mocha": "^3.1.2",
     "mock-aws-s3": "^0.2.1",
-    "nyc": "^6.4.4"
+    "nyc": "^6.4.4",
+    "tape": "^4.6.3"
   }
 }
